requirements.txt
streamlit>=1.33
videodb
google-generativeai
openai
groq
python-dotenv
